Update AME parser to use read_fwf
The format of the file is given in the header of the file. It's wrong,
but while finding that out I found that we can setup various lists and
dictionaries to then use read_fwf and parse the file in one go. We still
need to make a few adjustments once the parse is done, but they are
fairly minor.