Update AME mass table parsing test and fix bugs
Now we don't read the file line by line, the tests needed to be updated
to account for this. I also learnt about the pandas.testing module so
started to use that.

There was in a bug in the AME mass parsing. I assumed the atomic mass
always started with the same A as the isotope. This is not true so the
parsing was updated to read the value from the file. The atomic mass
error was also too large but is now scaled appropriately.