Update the top level mass table class
This is mainly adding in the new years and splitting the AME and NUBASE
parsing as they no longer have matching years. This also meant we have
to merge in a slightly different way and ensure we don't lose any data
unique to one set.

I have also removed the validation of the year. We currently have full
control over the years passed to the functions so should not get any
errors. The fact that AME and NUBASE now have different years means we
would have also had to add additional functionality so the decision was
made to delete. We can add back in if required.