Drop support for python 3.10
This is a bit of a cop-out, but tests are failing against this version
but not later ones and I don't want to spend time pushing changes and
using runners as my debug tool. 3.10 stopped receiving active support
over 2.5 years ago and will go EOL in 10 months so I'm comfortable
removing support. There are 4 newer versions people can use.